+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/executors/MkDirsCommandExecutor.kt
@@ -1,6 +1,7 @@
 package ru.hh.plugins.geminio.sdk.template.executors
 
 import com.android.tools.idea.wizard.template.RecipeExecutor
+import com.intellij.openapi.vfs.VirtualFile
 import ru.hh.plugins.extensions.EMPTY
 import ru.hh.plugins.geminio.sdk.recipe.models.commands.MkDirItem
 import ru.hh.plugins.geminio.sdk.recipe.models.commands.RecipeCommand
@@ -9,23 +10,31 @@ import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
 import java.io.File
 
 internal fun RecipeExecutor.execute(
+    targetDirectory: VirtualFile,
     command: RecipeCommand.MkDirs,
     executorData: GeminioRecipeExecutorData
 ) {
     for (rootItem in command.dirs) {
-        makeDirectories(rootItem, String.EMPTY, executorData)
+        makeDirectories(
+            targetDirectory = targetDirectory,
+            mkDirItem = rootItem,
+            combinedPath = String.EMPTY,
+            executorData = executorData
+        )
     }
 }
 
 private fun RecipeExecutor.makeDirectories(
+    targetDirectory: VirtualFile,
     mkDirItem: MkDirItem,
     combinedPath: String,
     executorData: GeminioRecipeExecutorData
 ) {
     val filePath = "$combinedPath/".takeIf { combinedPath.isNotEmpty() }.orEmpty() + requireNotNull(
         mkDirItem.name.evaluateString(
-            executorData.moduleTemplateData,
-            executorData.existingParametersMap
+            targetDirectory = targetDirectory,
+            moduleTemplateData = executorData.moduleTemplateData,
+            existingParametersMap = executorData.existingParametersMap
         )
     ) {
         "Recipe execution, 'mkDirs' command: Error with directory name evaluation [mkDirItem: $mkDirItem, combinedPath: $combinedPath, executorData: $executorData]"
@@ -33,6 +42,11 @@ private fun RecipeExecutor.makeDirectories(
     createDirectory(File(filePath))
 
     for (subdirectoryItem in mkDirItem.subDirs) {
-        makeDirectories(subdirectoryItem, filePath, executorData)
+        makeDirectories(
+            targetDirectory = targetDirectory,
+            mkDirItem = subdirectoryItem,
+            combinedPath = filePath,
+            executorData = executorData
+        )
     }
 }