+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/executors/OpenCommandExecutor.kt
@@ -1,6 +1,7 @@
 package ru.hh.plugins.geminio.sdk.template.executors
 
 import com.android.tools.idea.wizard.template.RecipeExecutor
+import com.intellij.openapi.vfs.VirtualFile
 import ru.hh.plugins.geminio.sdk.recipe.models.commands.RecipeCommand
 import ru.hh.plugins.geminio.sdk.template.mapping.expressions.evaluateString
 import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
@@ -8,10 +9,11 @@ import ru.hh.plugins.logger.HHLogger
 import java.io.File
 
 internal fun RecipeExecutor.execute(
+    targetDirectory: VirtualFile,
     command: RecipeCommand.Open,
     executorData: GeminioRecipeExecutorData
 ) = with(executorData) {
-    val filePath = command.file.evaluateString(moduleTemplateData, existingParametersMap)
+    val filePath = command.file.evaluateString(targetDirectory, moduleTemplateData, existingParametersMap)
 
     HHLogger.d("Open command [filePath: $filePath]")
     if (filePath == null) {