+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/executors/PredicateCommandExecutor.kt
@@ -1,12 +1,14 @@
 package ru.hh.plugins.geminio.sdk.template.executors
 
 import com.android.tools.idea.wizard.template.RecipeExecutor
+import com.intellij.openapi.vfs.VirtualFile
 import ru.hh.plugins.geminio.sdk.recipe.models.commands.RecipeCommand
 import ru.hh.plugins.geminio.sdk.template.mapping.expressions.evaluateBoolean
 import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
 import ru.hh.plugins.logger.HHLogger
 
 internal fun RecipeExecutor.execute(
+    targetDirectory: VirtualFile,
     command: RecipeCommand.Predicate,
     executorData: GeminioRecipeExecutorData
 ) = with(executorData) {
@@ -16,6 +18,7 @@ internal fun RecipeExecutor.execute(
     if (validIfExpression.evaluateBoolean(existingParametersMap)) {
         HHLogger.d("\tStart executing commands [validIf == true]")
         executeCommands(
+            targetDirectory = targetDirectory,
             commands = command.commands,
             executorData = executorData
         )
@@ -23,6 +26,7 @@ internal fun RecipeExecutor.execute(
         // Skip predicate command
         HHLogger.d("\tSkip commands execution [validIf == false], try to execute 'elseCommands' if exists")
         executeCommands(
+            targetDirectory = targetDirectory,
             commands = command.elseCommands,
             executorData = executorData
         )