+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/mapping/GeminioRecipeMapper.kt
@@ -3,6 +3,7 @@ package ru.hh.plugins.geminio.sdk.template.mapping
 import com.android.tools.idea.wizard.template.ModuleTemplateData
 import com.android.tools.idea.wizard.template.template
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
 import ru.hh.plugins.freemarker_wrapper.FreemarkerConfiguration
 import ru.hh.plugins.geminio.sdk.GeminioAdditionalParamsStore
 import ru.hh.plugins.geminio.sdk.GeminioSdkConstants
@@ -21,7 +22,7 @@ import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
  * Mapping from [ru.hh.plugins.geminio.sdk.recipe.models.GeminioRecipe]
  * into [ru.hh.plugins.geminio.sdk.models.GeminioTemplateData].
  */
-internal fun GeminioRecipe.toGeminioTemplateData(project: Project): GeminioTemplateData {
+internal fun GeminioRecipe.toGeminioTemplateData(project: Project, targetDirectory: VirtualFile): GeminioTemplateData {
     val geminioRecipe = this
 
     val existingParametersMap = mutableMapOf<String, AndroidStudioTemplateParameter>()
@@ -36,6 +37,7 @@ internal fun GeminioRecipe.toGeminioTemplateData(project: Project): GeminioTempl
         recipe = { templateData ->
             val moduleTemplateData = templateData as ModuleTemplateData
             executeGeminioRecipe(
+                targetDirectory = targetDirectory,
                 geminioRecipe = geminioRecipe,
                 executorData = GeminioRecipeExecutorData(
                     project = project,
@@ -45,7 +47,7 @@ internal fun GeminioRecipe.toGeminioTemplateData(project: Project): GeminioTempl
                     resolvedParamsMap = existingParametersMap.asIterable().associate { entry ->
                         entry.key to entry.value.value
                     }.plus(
-                        getHardcodedParamsMap(moduleTemplateData, existingParametersMap)
+                        getHardcodedParamsMap(targetDirectory, moduleTemplateData, existingParametersMap)
                     ).plus(
                         paramsStore
                     ),
@@ -69,11 +71,12 @@ internal fun GeminioRecipe.toGeminioTemplateData(project: Project): GeminioTempl
 }
 
 private fun getHardcodedParamsMap(
+    targetDirectory: VirtualFile,
     moduleTemplateData: ModuleTemplateData,
     existingParametersMap: Map<String, AndroidStudioTemplateParameter>
 ): Map<String, Any?> {
     val packageNameParameter = existingParametersMap[GeminioSdkConstants.FEATURE_PACKAGE_NAME_PARAMETER_ID]
-        as? AndroidStudioTemplateStringParameter
+            as? AndroidStudioTemplateStringParameter
 
     val (packageName, applicationPackage) = when {
         packageNameParameter != null -> {
@@ -85,9 +88,15 @@ private fun getHardcodedParamsMap(
         }
     }
 
+    var uiTestPackageName = targetDirectory.path.replace("/", ".")
+    uiTestPackageName = uiTestPackageName.substring(
+        uiTestPackageName.indexOf(packageName)
+    )
+
     return mapOf(
         HardcodedParams.PACKAGE_NAME to packageName,
         HardcodedParams.APPLICATION_PACKAGE to applicationPackage,
+        HardcodedParams.CURRENT_DIR_PACKAGE_NAME to uiTestPackageName
     )
 }
 
@@ -101,4 +110,10 @@ private object HardcodedParams {
      * Package name from current gradle module.
      */
     const val APPLICATION_PACKAGE = "applicationPackage"
+
+    /**
+     * Package name from current gradle module.
+     */
+    const val CURRENT_DIR_PACKAGE_NAME = "currentDirPackageName"
+
 }